% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt}
\alias{fmt}
\title{Assign or Retrieve Data Format}
\usage{
fmt(x, ...)
}
\arguments{
\item{x}{Target object.}

\item{...}{Other arguments (not used).}

\item{y}{Character string(s) specifying the name(s) of the variables or attributes to be assigned data formats.}

\item{value}{Character string(s) specifying the data format to be assigned.}
}
\value{
Object with a format \code{fmt} attribute attached to it.
}
\description{
Functions to assigns or retrieve data formatting information to and from an object.
}
\details{
Named argiments in the format that are not in the target object are ignored.
         The \code{fmt} attribute may be erased by assigning a \code{NULL} or empty 
         string \code{""} value to it.
}
\section{Functions}{

\describe{
  \item{\code{fmt}}{Generic \code{fmt} method.}
  \item{\code{fmt.default}}{Default \code{fmt} method.}
  \item{\code{fmt<-}}{Generic format assignment method.}
  \item{\code{fmt<-.default}}{Default format assignment method. See Examples for usage.} 
}
}

\examples{
x <- data.frame(year = 2010:2014, month = 6, day = 1:5, measurement = rnorm(5))
fmt(x, "year")  <- "YYYY"
fmt(x, "month") <- "MM"
fmt(x, "day")   <- "DD"

# Equivalently:
fmt(x) <- c(year = "YYYY", month = "MM", day = "DD")
fmt(x, c("year", "month", "day")) <- c("YYYY", "MM", "DD")

# Show attributes:
attributes(x)
fmt(x)

# Erase 'fmt' attribute:
fmt(x) <- NULL

}
\seealso{
\code{\link{metadata}}
}
