% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\title{File Header Information}
\usage{
header(x, ...)
}
\arguments{
\item{x}{File name or object.}
}
\description{
Extracts header information from certain data files.
}
\section{Functions}{

\describe{
  \item{\code{header}}{Default \code{header} method.}
  \item{\code{header.default}}{Default \code{header} retrieval method.}
  \item{\code{header<-}}{Generic \code{header} assignment method. See Examples for usage.}
  \item{\code{header<-.default}}{Default\code{header} assignment method. See Examples for usage.} 
}
}

\examples{

}
\seealso{
\code{\link[gulf.data]{minilog}}
}
