% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.R
\name{keyword}
\alias{keyword}
\title{Keyword Functions}
\usage{
keyword(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Other arguments (not used).}

\item{value}{Character string(s) specifying the index key to be assigned.}
}
\description{
Functions to assign, retrieve or look up object keyword attributes.
}
\section{Functions}{

\describe{
  \item{\code{keyword}}{Default \code{key} method.}
  \item{\code{keyword.default}}{Default \code{format} retrieval method. If no \code{key} attribute is 
                            defined, then a call to \code{\link[base]{format}} is made.}
  \item{\code{key<-}}{Generic \code{key} assignment method. See Examples for usage.}
  \item{\code{key<-.default}}{Default\code{key} assignment method. See Examples for usage.} 
}
}

\examples{
# Build sample data:
x <- data.frame(year = 1990:2014, measurement = rpois(25, 5))
keyword(x) <- c("Annual", "survey") # Assign keywords
keyword(x) # Retrieve key.

}
\seealso{
\code{\link{metadata}}, \code{\link{description}}, \code{\link{units}}, \code{\link[gulf.metadata]{format}}
}
