% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Retrieve Metadata}
\usage{
metadata(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Other arguments (not used).}
}
\description{
Function to retrieve metadata from an object.
}
\section{Methods}{

\describe{
  \item{\code{\link{key}}}{Functions to assign, retrieve or check an index key for a given object.}
  \item{\code{\link{description}}}{Assign or retrieve a text description to an object.}
  \item{\code{\link{units}}}{Assigns or retrieves measurement units assigned to a data object.}
  \item{\code{\link{fmt}}}{Functions to assigns or retrieve data formatting information to and from an object.}
  \item{\code{\link{keyword}}}{Functions to assign, retrieve or look up object keyword attributes.}
  \item{\code{\link{project}}}{Functions to retrieve or assign project identifiers to objects.}   
  \item{\code{\link{header}}}{Function to assign and retrieve file header information associated with a data object.} 
}
}

\examples{
# Define sample data set:
x <- data.frame(year = 2010:2014,
                site = letters[1:5],
                weight = exp(rnorm(5)),
                number = rpois(5,5))

# Define metadata attributes:
units(x, "weight") <- "kilogram"
description(x) <- c(year = "Year of sampling",
                    site = "Sampling site",
                    weight = "Weight of specimen",
                    number = "Number observed")
format(x, "year") <- "YYYY"
key(x) <- "year"
keywords(x) <- c("annual", "survey")

# Retrieve metadata:
metadata(x)

}
