% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project Identifiers}
\usage{
project(x, ...)
}
\arguments{
\item{x}{Target object.}

\item{...}{Other arguments (not used).}

\item{value}{Character string(s) specifying the project identifier to be assigned.}

\item{verbose}{Logical value specifying whether to return a character string result in long form.}
}
\value{
An object with a \code{project} attribute attached to it.
}
\description{
Functions to retrieve or assign project identifiers to objects.
}
\details{
The \code{project} attribute may be erased by assigning a \code{NULL} value to it.
}
\section{Functions}{

\describe{
  \item{\code{project}}{Generic \code{project}' extraction method.}
  \item{\code{project.default}}{Default \code{project} extraction method.} 
  \item{\code{project.character}}{Find project name.} 
  \item{\code{project<-}}{Generic \code{project} assignment method.}
  \item{\code{project<-.default}}{Default \code{project} assignment method.}
}
}

\examples{
x <- data.frame(year = 2010:2014, measurement = rnorm(5))
project(x) <- "test.survey"

# Show attributes:
attributes(x)

# Erase 'project' attribute:
project(x) <- NULL   # Remove all projects.

# Load entire project table:
project()

# Find project name in project table:
project("sc")
project("snow crab survey")
project("snow crab fecundity")
project("rv")
project("september survey") 
project("Northumberland") 
project("rv", verbose = TRUE) # Long form.

}
\seealso{
\code{\link{metadata}}, \code{\link{key}}, \code{\link{description}}, \code{\link{units}}, \code{\link[gulf.metadata]{format}}
}
